pyproject.toml
[project]
name = "litellm-firewall"
version = "0.1.0"
description = "A minimal LLM firewall demonstration using LiteLLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm[proxy]>=1.82.0",
    "openai>=2.26.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
]
